+++ Insight/database/db_tables/discord/discord_channels.py
@@ -125,7 +125,8 @@ class Channels(dec_Base.Base,discord_channel_base):
             return return_str
         else:
             return "This radar feed has no synced contact tokens. Contact tokens block your allies from " \
-                   "appearing as potential targets in a radar feed. You can add a token by running the command '!sync'."
+                   "appearing as potential targets in a radar or proximity feed. You can add a token by running the " \
+                   "'!sync' command. \nNote: This message will only appear once."
 
     def sync_api_contacts(self, service_module):
         db: Session = service_module.get_session()