+++ Insight/discord_bot/channel_types/FiltersVisualsEmbedded/__init__.py
@@ -6,6 +6,6 @@ from .VisualEnfeedCompact import VisualEnfeedCompact
 from .VisualCapRadarCompact import VisualCapRadarCompact
 from .VisualEnfeedLinkOnly import VisualEnfeedLinkOnly
 from .VisualCapRadarLinkOnly import VisualCapRadarLinkOnly
-from .VisualCapRadarFunctional import VisualCapRadarFunctional
+from .VisualCapRadarUtility import VisualCapRadarUtility
 from .VisualProximityWatch import VisualProximityWatch