+++ Insight/discord_bot/channel_types/ProximityWatch.py
@@ -15,7 +15,10 @@ class ProximityWatch(capRadar):
 
     @classmethod
     def get_template_desc(cls):
-        return "Proximity Watch - Null"
+        return "Proximity Watch - Tracks any hostile fleet activity near your base systems or in a selection of " \
+               "systems, constellations, or regions. Proximity watches are ideal for finding potential fleets to " \
+               "fight, tracking hostile fleet movement within your region, or altering you of nearby hostiles" \
+               " within a few jumps of your base systems."
 
     def __str__(self):
         return "Proximity Watch"