+++ Insight/discord_bot/channel_types/TemplateFeeds/ExcavatorLosses.py
@@ -25,7 +25,7 @@ class ExcavatorLosses(enFeed):
 
     @classmethod
     def get_template_desc(cls):
-        return "Excavator Losses - Displays all excavator mining drone losses."
+        return "Excavator Losses - Displays excavator mining drone losses."
 
     def __str__(self):
         return "Excavator Feed"