+++ Insight/discord_bot/channel_types/TemplateFeeds/__init__.py
@@ -1,9 +1,10 @@
 from .SuperLosses import SuperLosses
-from .AbyssalLosses import AbyssalLosses
+from .CapitalLosses import CapitalLosses
+from .BigKills import BigKills
+from .Freighters import Freighters
 from .ExcavatorLosses import ExcavatorLosses
+from .AbyssalLosses import AbyssalLosses
 from .OfficerHunter import OfficerHunter
 from .UniverseSupers import UniverseSupers
 from .RareShips import RareShips
-from .Freighters import Freighters
 # from .AllianceTournamentStream import AllianceTournamentStreamSystem # remove until next year
-from .CapitalLosses import CapitalLosses